feat: SQLAlchemy imperative mapping style 기본 구조 구현 (main)
excercieses/excercies01 폴더 생성 및 기본 구조 구현

config 폴더 생성
- client.py : DB 세션 클라이언트 생성
- start_mappers.py : ORM Mapper 시작 및 등록
- sa_models.py : SQLAlchemy 모델 정의
- sqls : SQL 파일

- 모델 설계 및 Mapper 적용
- excercieses 폴더 생성 및 기본 구조 구현

+++ usages/sqlalchemy_imperative_mapping_style/config/sa_models.py
@@ -1,7 +1,6 @@
-# coding: utf-8
 from sqlalchemy import CHAR, Column, String
-from sqlalchemy.dialects.mysql import INTEGER
-from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.dialects.mysql import INTEGER, TINYINT
+from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
 metadata = Base.metadata
@@ -10,10 +9,10 @@ metadata = Base.metadata
 class Member(Base):
     __tablename__ = 'member'
 
-    pk = Column(INTEGER(1), primary_key=True)
+    pk = Column(INTEGER(10), primary_key=True)
     nanoid = Column(CHAR(24), nullable=False)
     name = Column(String(32))
-    age = Column(INTEGER(1))
+    age = Column(TINYINT(3))
     address1 = Column(String(1024))
     address2 = Column(String(1024))
